+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * System metrics are now reported for first-class `ViewLoad` spans
   [#475](https://github.com/bugsnag/bugsnag-android-performance/pull/475)
+  * Enable runtime control of fragment instrumentation via a new `FragmentInstrumentation.enabled` flag
+  [#490](https://github.com/bugsnag/bugsnag-android-performance/pull/490)
 
 ### Bug fixes